[JE] Heads Up Display and Sound
+ Added frame around game
+ Added coin counter to HUD
+ Added animated coin on HUD
+ Added toggleable debug stats to HUD
+ Added function to toggle HUD on or off (broken)
* Sounds now pre-loaded when the level loads (fixes coin pickup lag)

+++ ShapeSprint/CGameMap.cpp
@@ -5,7 +5,7 @@ CGameMap::CGameMap()
 }
 
 // NEEDS COMMENTING //
-bool CGameMap::LoadTheMap(FullLevel& map, float startCoods[], float checkpointCoords[], float endCoords[], float& timeLimit, float& mapWidth, float mapHeight, string levelName)
+bool CGameMap::LoadTheMap(FullLevel& map, float startCoods[], float checkpointCoords[], float endCoords[], int& timeLimit, int& mapWidth, float mapHeight, string levelName)
 {
 	ifstream levelFile; //input file - Should be variations of level[x].txt
